update randao wrapper contract

+++ contracts/interface/IRandaoWrapper.sol
@@ -2,15 +2,7 @@
 pragma solidity ^0.8.9;
 
 interface IRandaoWrapper {
-    event LogRequestRandom(
-        uint256 indexed requestId,
-        address indexed from,
-        uint256 indexed bnum,
-        uint256 bounty,
-        uint256 deposit,
-        uint16 commitBalkline,
-        uint16 commitDeadline
-    );
+    event LogRequestRandom(uint256 indexed requestId, address indexed from, uint256 indexed bnum, uint256 bounty, uint256 deposit, uint256 commitBalkline, uint256 commitDeadline);
 
     event LogRawFulfillRandomWords(uint256, uint256[]);
 
@@ -20,43 +12,15 @@ interface IRandaoWrapper {
     /// @dev Send a request for random numbers generation.
     /// @param numWords for the number of request random numbers.
     /// @return requestId for request random numbers.
-    function requestRandomWords(
-        bytes32,
-        uint64,
-        uint16,
-        uint32,
-        uint32 numWords
-    ) external returns (uint256 requestId);
+    function requestRandomWords(bytes32, uint256, uint256, uint256, uint256 numWords) external returns (uint256 requestId);
 
     /// @dev Random generation response callback after generating successfully. must be provided by you own contract.
     /// @param requestId for request random numbers.
     function rawFulfillRandomWords(uint256 requestId) external;
 
-    function latestRoundData()
-        external
-        view
-        returns (
-            uint80 roundId,
-            int256 answer,
-            uint256 startedAt,
-            uint256 updatedAt,
-            uint80 answeredInRound
-        );
+    function latestRoundData() external view returns (uint256 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint256 answeredInRound);
 
-    function getFeeConfig()
-        external
-        view
-        returns (
-            uint32,
-            uint32,
-            uint32,
-            uint32,
-            uint32,
-            uint24,
-            uint24,
-            uint24,
-            uint24
-        );
+    function getFeeConfig() external view returns (uint256, uint256, uint256, uint256, uint256, uint256, uint256, uint256, uint256);
 
     function refundFee(uint256 requestId) external payable;
 }
